Add BigVector and VectorView.

+++ FParLin/const.hpp
@@ -8,38 +8,38 @@
 
 using namespace std;
 
-using Const_ = boost::variant<int, vector<int>>;
+using Const_ = boost::variant<double, vector<double>>;
 
 struct Const : Const_ {
 	Const() : Const_{ 0 } {}
-	Const(int a) : Const_(a) {}
-	Const(vector<int> a) : Const_(a) {}
+	Const(double a) : Const_(a) {}
+	Const(vector<double> a) : Const_(a) {}
 
-	explicit operator int() {
-		return boost::get<int>(*this);
+	explicit operator double() {
+		return boost::get<double>(*this);
 	}
 
-	explicit operator vector<int>() {
-		return boost::get<vector<int>>(*this);
+	explicit operator vector<double>() {
+		return boost::get<vector<double>>(*this);
 	}
 };
 
 Const operator+(const Const& a, const Const& b) {
-	return boost::get<int>(a) + boost::get<int>(b);
+	return boost::get<double>(a) + boost::get<double>(b);
 }
 
 Const operator*(const Const& a, const Const& b) {
-	return boost::get<int>(a) * boost::get<int>(b);
+	return boost::get<double>(a) * boost::get<double>(b);
 }
 
 struct const_printer : boost::static_visitor<void> {
 	const_printer(ostream& out) : out{ out } {}
 
-	void operator()(int i) const {
+	void operator()(double i) const {
 		out << i;
 	}
 
-	void operator()(const vector<int>& v) const {
+	void operator()(const vector<double>& v) const {
 		out << "[";
 		for (unsigned i = 0; i < v.size(); ++i) {
 			out << (i ? " " : "") << v[i];
